bni-digivouch/source/ayopop-proxy/app/requirements.txt
aiohttp
requests
PyJWT